Inject tenant in all* logging records
The ContextVar is not always passed down properly, we'll address
them case by case.
